share code path with the non-mtp batch decoder
