+++ tensorrt_llm/_torch/pyexecutor/decoder.py
@@ -23,6 +23,8 @@ from tensorrt_llm.mapping import Mapping
 from .llm_request import LlmRequest, LlmRequestState
 from .scheduler import ScheduledRequests
 
+from .b10 import B10Decoder
+
 
 @dataclass
 class DecoderState:
@@ -175,6 +177,8 @@ class TorchDecoder(Decoder):
     def __init__(self, max_seq_len: int, mixed_decoder: bool = False):
         self.max_seq_len = max_seq_len
         self.mixed_decoder = mixed_decoder
+
+        torch.manual_seed(42)
 
     def setup_decoder(self, scheduled_requests: ScheduledRequests,
                       model_outputs):
@@ -334,6 +338,15 @@ class TorchDecoder(Decoder):
                       model_outputs) -> DecoderState:
         logits = model_outputs["logits"]
         new_tokens_device = torch.argmax(logits, dim=-1)
+
+        # BASETEN DECODING BEGIN
+        _, _, sampled_tokens = B10Decoder.custom_decode(scheduled_requests,
+                                                                  model_outputs,
+                                                                  process_all_requests=True)
+        assert len(new_tokens_device) == len(sampled_tokens)
+        new_tokens_device[:] = sampled_tokens
+        # BASETEN DECODING END
+
         new_tokens_host = new_tokens_device.to('cpu', non_blocking=True)
         decoder_event = torch.cuda.Event()
         decoder_event.record()
